Start renaming files and updating references
