+++ qiskit/classicalregister.py
