+++ qiskit/extensions/simulator/wait.py
@@ -12,8 +12,8 @@ Aer's qasm_simulator single qubit wait gate.
 """
 from qiskit import Gate
 from qiskit import QuantumCircuit
-from qiskit._instructionset import InstructionSet
-from qiskit._quantumregister import QuantumRegister
+from qiskit.instructionset import InstructionSet
+from qiskit.quantumregister import QuantumRegister
 from qiskit.qasm import _node as node
