+++ qiskit/instruction.py
@@ -26,9 +26,9 @@ Instructions are identified by the following fields, and are serialized as such 
 """
 import sympy
 
-from ._qiskiterror import QISKitError
-from ._quantumregister import QuantumRegister
-from ._classicalregister import ClassicalRegister
+from .qiskiterror import QISKitError
+from .quantumregister import QuantumRegister
+from .classicalregister import ClassicalRegister
 
 
 class Instruction(object):
