+++ qiskit/instructionset.py
@@ -8,8 +8,8 @@
 """
 Instruction collection.
 """
-from ._instruction import Instruction
-from ._qiskiterror import QISKitError
+from .instruction import Instruction
+from .qiskiterror import QISKitError
 
 
 class InstructionSet(object):
