+++ qiskit/measure.py
@@ -8,11 +8,11 @@
 """
 Quantum measurement in the computational basis.
 """
-from ._instruction import Instruction
-from ._instructionset import InstructionSet
-from ._quantumcircuit import QuantumCircuit
-from ._quantumregister import QuantumRegister
-from ._classicalregister import ClassicalRegister
+from .instruction import Instruction
+from .instructionset import InstructionSet
+from .quantumcircuit import QuantumCircuit
+from .quantumregister import QuantumRegister
+from .classicalregister import ClassicalRegister
 
 
 class Measure(Instruction):
