+++ qiskit/qasm/_node/__init__.py
@@ -14,7 +14,7 @@ from ._creg import Creg
 from ._customunitary import CustomUnitary
 from ._expressionlist import ExpressionList
 from ._external import External
-from ._gate import Gate
+from .gate import Gate
 from ._gatebody import GateBody
 from ._id import Id
 from ._idlist import IdList
@@ -22,14 +22,14 @@ from ._if import If
 from ._indexedid import IndexedId
 from ._intnode import Int
 from ._format import Format
-from ._measure import Measure
+from .measure import Measure
 from ._opaque import Opaque
 from ._prefix import Prefix
 from ._primarylist import PrimaryList
 from ._program import Program
 from ._qreg import Qreg
 from ._real import Real
-from ._reset import Reset
+from .reset import Reset
 from ._unaryoperator import UnaryOperator
 from ._universalunitary import UniversalUnitary
 from ._nodeexception import NodeException
