+++ qiskit/qiskiterror.py
