+++ qiskit/schema_validation.py
