+++ qiskit/tools/file_io.py
@@ -16,7 +16,7 @@ import numpy
 from sympy import Basic
 
 from qiskit.result._utils import result_from_old_style_dict
-from qiskit._qiskiterror import QISKitError
+from qiskit.qiskiterror import QISKitError
 from qiskit.backends import BaseBackend
